Splash Screen, Start Up Screen, Scene Renaming
Ok, so I added two new scenes - SplashScene and StartUpScene.  Also, I
renamed our "New Scene" (which was our main scene) to "Main Scene", and
made a Scenes folder for organizational purposes...

See "Updates Splash and Start" email for more info.
SplashScreen is the first scene that starts on gameplay, it's simple and
just has a timer attached with a background texture with metro pollution
on it...I'm not an artist so if you guys want to change the background
texture feel free.  After the timer stops, it calls a co routine which
loads a new start up screen, which when clicking play directs the user
to our MainScene...It takes a bit to load the MainScene, so I might try
to put in a loading screen tomorrow...some of the code is already in
place, it just needs to be fixed --